+++ src/frontend/components/actionItemDisplay.tsx
@@ -28,11 +28,12 @@ export interface ActionItemDisplayProps {
   nonHiddenWorkItemTypes: WorkItemType[];
   allWorkItemTypes: WorkItemType[];
   allowAddNewActionItem: boolean;
+  shouldFocusActionItems?: boolean;
 
   onUpdateActionItem: (feedbackItemId: IFeedbackItemDocument) => void;
 }
 
-export const ActionItemDisplay: React.FC<ActionItemDisplayProps> = ({ feedbackItemId, feedbackItemTitle, team, boardId, boardTitle, defaultIteration, defaultAreaPath, actionItems, nonHiddenWorkItemTypes, allWorkItemTypes, allowAddNewActionItem, onUpdateActionItem }) => {
+export const ActionItemDisplay: React.FC<ActionItemDisplayProps> = ({ feedbackItemId, feedbackItemTitle, team, boardId, boardTitle, defaultIteration, defaultAreaPath, actionItems, nonHiddenWorkItemTypes, allWorkItemTypes, allowAddNewActionItem, shouldFocusActionItems = true, onUpdateActionItem }) => {
   const trackActivity = useTrackMetric(reactPlugin, "ActionItemDisplay");
 
   const [isLinkedWorkItemLoaded, setIsLinkedWorkItemLoaded] = useState(false);
@@ -145,14 +146,20 @@ export const ActionItemDisplay: React.FC<ActionItemDisplayProps> = ({ feedbackIt
     setLinkedWorkItem(workItem[0] ? workItem[0] : null);
   }, []);
 
+  const closeLinkExistingWorkItemDialog = useCallback((event?: React.SyntheticEvent) => {
+    event?.stopPropagation();
+
+    linkExistingWorkItemDialogRef.current!.close();
+  }, []);
+
   const linkExistingWorkItem = useCallback(async () => {
     const updatedFeedbackItem = await itemDataService.addAssociatedActionItem(boardId, feedbackItemId, linkedWorkItem!.id);
 
     appInsights.trackEvent({ name: TelemetryEvents.ExistingWorkItemLinked, properties: { workItemTypeName: linkedWorkItem!.fields["System.WorkItemType"] } });
 
+    closeLinkExistingWorkItemDialog();
     onUpdateActionItem(updatedFeedbackItem);
-    linkExistingWorkItemDialogRef.current!.close();
-  }, [linkedWorkItem, boardId, feedbackItemId, onUpdateActionItem]);
+  }, [linkedWorkItem, boardId, feedbackItemId, closeLinkExistingWorkItemDialog, onUpdateActionItem]);
 
   const handleLinkExistingWorkItemClick = useCallback(
     (mouseEvent: React.MouseEvent<HTMLAnchorElement | HTMLButtonElement | HTMLDivElement> | undefined = undefined) => {
@@ -169,9 +176,9 @@ export const ActionItemDisplay: React.FC<ActionItemDisplayProps> = ({ feedbackIt
 
   const renderWorkItemCard = useCallback(
     (item: WorkItem, areActionIconsHidden: boolean) => {
-      return <ActionItem key={item.id} feedbackItemId={feedbackItemId} boardId={boardId} actionItem={item} nonHiddenWorkItemTypes={nonHiddenWorkItemTypes} allWorkItemTypes={allWorkItemTypes} onUpdateActionItem={onUpdateActionItem} areActionIconsHidden={areActionIconsHidden} shouldFocus={!initialRender} />;
+      return <ActionItem key={item.id} feedbackItemId={feedbackItemId} boardId={boardId} actionItem={item} nonHiddenWorkItemTypes={nonHiddenWorkItemTypes} allWorkItemTypes={allWorkItemTypes} onUpdateActionItem={onUpdateActionItem} areActionIconsHidden={areActionIconsHidden} shouldFocus={!initialRender && shouldFocusActionItems} />;
     },
-    [feedbackItemId, boardId, nonHiddenWorkItemTypes, allWorkItemTypes, onUpdateActionItem, initialRender],
+    [feedbackItemId, boardId, nonHiddenWorkItemTypes, allWorkItemTypes, onUpdateActionItem, initialRender, shouldFocusActionItems],
   );
 
   const renderAllWorkItemCards = useCallback(() => {
@@ -227,19 +234,19 @@ export const ActionItemDisplay: React.FC<ActionItemDisplayProps> = ({ feedbackIt
         </div>
       )}
       {renderAllWorkItemCards()}
-      <dialog ref={linkExistingWorkItemDialogRef} className="link-existing-work-item-dialog dialog-width-sm" aria-label="Link existing work item" onClose={() => linkExistingWorkItemDialogRef.current!.close()}>
+      <dialog ref={linkExistingWorkItemDialogRef} className="link-existing-work-item-dialog dialog-width-sm" aria-label="Link existing work item" onCancel={event => event.stopPropagation()} onClose={event => event.stopPropagation()}>
         <div className="header">
           <h2 className="title">{t("common_link_existing_work_item")}</h2>
-          <button onClick={() => linkExistingWorkItemDialogRef.current!.close()} aria-label="Close">
+          <button onClick={closeLinkExistingWorkItemDialog} aria-label="Close">
             {getIconElement("close")}
           </button>
         </div>
         <div className="subText">
           <div className="form-group">
-            <input className="search-box" placeholder="Enter the exact work item id" role="searchbox" onChange={handleInputChange}></input>
+            <input className="search-box" placeholder="Enter the exact work item ID" role="searchbox" onChange={handleInputChange}></input>
             <div className="output-container">
               {isLinkedWorkItemLoaded && linkedWorkItem && renderWorkItemCard(linkedWorkItem, true)}
-              {isLinkedWorkItemLoaded && !linkedWorkItem && <div className="work-item-not-found">The work item you are looking for was not found. Please verify the id.</div>}
+              {isLinkedWorkItemLoaded && !linkedWorkItem && <div className="work-item-not-found">The work item you are looking for was not found. Please verify the ID.</div>}
             </div>
           </div>
         </div>
@@ -247,7 +254,7 @@ export const ActionItemDisplay: React.FC<ActionItemDisplayProps> = ({ feedbackIt
           <button className="button" disabled={!linkedWorkItem} onClick={linkExistingWorkItem}>
             {t("common_link_work_item")}
           </button>
-          <button className="default button" onClick={() => linkExistingWorkItemDialogRef.current!.close()}>
+          <button className="default button" onClick={closeLinkExistingWorkItemDialog}>
             {t("common_cancel")}
           </button>
         </div>